+++ NEWS.md
@@ -3,3 +3,8 @@
 * Added ody_summarise_df
 * Added ody_define_timepoints
 * Added ody_pdx_model_sensitivity
+
+# odytools 0.2.0
+
+* Added data quality functions ody_verify_completeness and ody_verify_conformance.
+* Added ody_render_quality_report to render a html with the result of ody_verify_completeness and ody_verify_conformance.